Fix non-ffmpeg screen gen
Non-ffmpeg screen gen was creating files with the prefix 001, 002, 003, (basically which frame it was in order) while ffmpeg and the rest of the script was expecting frame numbers from the video like 783, 1835, 2003.

The script wasn't checking whether ffmpeg was actually in path. It now switched to non-ffmpeg screengen if it isn't found.

Also no longer check if a file's group name was previously analysed if it doesn't have a group name.